feat(zink): move u256 ffi to ext
