+++ zink/src/primitives/mod.rs
@@ -2,7 +2,6 @@
 
 pub mod address;
 pub mod bytes;
-pub mod numeric;
 pub mod u256;
 
 pub use {address::Address, bytes::*, u256::U256};
