+++ zink/src/primitives/u256.rs
@@ -4,6 +4,8 @@ use crate::{asm, primitives::Bytes32, storage::Value};
 use core::ops::Sub;
 
 /// Account address
+///
+/// TODO: impl Numeric trait for U256
 #[repr(C)]
 #[derive(Clone, Copy, PartialEq, PartialOrd, Debug)]
 pub struct U256(Bytes32);
@@ -17,13 +19,13 @@ impl U256 {
     /// u256 add
     #[inline(always)]
     pub fn add(self, other: Self) -> Self {
-        unsafe { asm::u256_add(self, other) }
+        unsafe { asm::ext::u256_add(self, other) }
     }
 
     /// u256 less than
     #[inline(always)]
     pub fn lt(self, other: Self) -> bool {
-        unsafe { asm::u256_lt(other, self) }
+        unsafe { asm::ext::u256_lt(other, self) }
     }
 
     /// u256 eq
@@ -35,19 +37,19 @@ impl U256 {
     /// u256 sub
     #[inline(always)]
     pub fn sub(self, other: Self) -> Self {
-        unsafe { asm::u256_sub(other, self) }
+        unsafe { asm::ext::u256_sub(other, self) }
     }
 
     /// u256 div
     #[inline(always)]
     pub fn div(self, other: Self) -> Self {
-        unsafe { asm::u256_div(self, other) }
+        unsafe { asm::ext::u256_div(self, other) }
     }
 
     /// max of u256
     #[inline(always)]
     pub fn max() -> Self {
-        unsafe { asm::u256_max() }
+        unsafe { asm::ext::u256_max() }
     }
 
     pub fn to_bytes32(&self) -> Bytes32 {
@@ -61,23 +63,22 @@ impl U256 {
 
     #[inline(always)]
     pub fn addmod(self, other: Self, modulus: Self) -> Self {
-        unsafe { asm::u256_addmod(modulus, other, self) }
+        unsafe { asm::ext::u256_addmod(modulus, other, self) }
     }
 
     /// Mulmod for U256
     #[inline(always)]
     pub fn mulmod(self, other: Self, modulus: Self) -> Self {
-        unsafe { asm::u256_mulmod(modulus, other, self) }
+        unsafe { asm::ext::u256_mulmod(modulus, other, self) }
     }
 }
 
 impl Sub for U256 {
     type Output = Self;
 
-    /// u256 sub
     #[inline(always)]
     fn sub(self, other: Self) -> Self::Output {
-        unsafe { asm::u256_sub(self, other) }
+        unsafe { asm::ext::u256_sub(self, other) }
     }
 }
